removed hibernate from project and implemented wallet-history query in jpa native query

+++ src/main/java/com/example/virtualwallet/models/dtos/pageable/WalletPageOutput.java
@@ -24,5 +24,5 @@ public class WalletPageOutput {
 
     private int historyPages;
 
-    private int historySize;
+    private long historySize;
 }
